# src/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go + React + AI Platform API
  description: 深度研究与报告生成平台 API
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/ileego/go_react_ai
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
paths:
...
  /search/reports:
    get:
      summary: 全文搜索报告
      operationId: searchReports
      tags:
        - Search
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: 搜索关键词
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
        "400":
          description: 缺少搜索关键词
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /files/presigned-upload:
    get:
      summary: 获取客户端直传的预签名上传链接
      operationId: presignedUpload
      tags:
        - Files
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: 文件名
        - name: content_type
          in: query
          required: true
          schema:
            type: string
          description: 文件 MIME 类型
      responses:
        "200":
          description: 预签名上传链接
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUploadResponse"
        "400":
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        topic:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        content:
          type: string
        sources:
          type: array
          items:
            type: string
        created_by:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ReportListResponse:
      type: object
      properties:
        code:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
    File:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        content_type:
          type: string
        size:
          type: integer
          format: int64
        bucket:
          type: string
        created_at:
          type: string
          format: date-time
...
    PresignedUploadResponse:
      type: object
      properties:
        code:
          type: integer
        data:
          type: object
          properties:
            upload_url:
              type: string
            file:
              $ref: "#/components/schemas/File"
            expires_in:
              type: integer
              format: int64
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT